% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek1.R
\name{answerWeek1}
\alias{answerWeek1}
\title{Week 1 Answer Function}
\usage{
answerWeek1(data)
}
\arguments{
\item{data}{The randomly generated question data}
}
\description{
This function computes the answer to Week 1's practice script
}
