% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek2a.R
\name{answerWeek2a}
\alias{answerWeek2a}
\title{Week 2a Answer Function}
\usage{
answerWeek2a(data)
}
\arguments{
\item{data}{The randomly generated question data}
}
\description{
This function computes the answer to the first question of Week 2's practice script
}
