% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek2b.R
\name{answerWeek2b}
\alias{answerWeek2b}
\title{Week 2b Answer Function}
\usage{
answerWeek2b(data, newM, newSD)
}
\arguments{
\item{data}{The randomly generated question data}

\item{newM}{The randomly generated new sample mean}

\item{newSd}{The randomly generated new sample SD}
}
\description{
This function computes the answer to the second question of Week 2's practice script
}
