% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek3.R
\name{answerWeek3}
\alias{answerWeek3}
\title{Week 3 Answer Function}
\usage{
answerWeek3(data, popMean, popSD)
}
\arguments{
\item{data}{The randomly generated question data}

\item{popMean}{The randomly generated population mean}

\item{popSD}{The randomly generated population SD}
}
\description{
This function computes the answers to Week 3's practice script
}
