% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek4.R
\name{answerWeek4}
\alias{answerWeek4}
\title{Week 4 Answer Function}
\usage{
answerWeek4(data, popMean)
}
\arguments{
\item{data}{The randomly generated question data}

\item{popMean}{The randomly generated population mean}
}
\description{
This function computes the answers to Week 4's practice script
}
