% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek5.R
\name{answerWeek5}
\alias{answerWeek5}
\title{Week 5 Answer Function}
\usage{
answerWeek5(sample1, sample2, type)
}
\arguments{
\item{sample1}{The randomly generated data for sample 1}

\item{sample2}{The randomly generated data for sample 1}

\item{type}{The type of t-test you want to conduct}
}
\description{
This function computes the answers to Week 5's practice script
}
