% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek6.R
\name{answerWeek6}
\alias{answerWeek6}
\title{Week 6 Answer Function}
\usage{
answerWeek6(data)
}
\arguments{
\item{data}{The randomly generated problem data}
}
\description{
This function computes the answers to Week 5's practice script
}
