% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek7b.R
\name{answerWeek7b}
\alias{answerWeek7b}
\title{Week 7b Answer Function}
\usage{
answerWeek7b(d, n)
}
\arguments{
\item{d}{The randomly generated Cohen's d}

\item{n}{The randomly generated sample size}
}
\description{
This function computes the answer to the second question of Week 7's practice script
}
