% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek8a.R
\name{answerWeek8a}
\alias{answerWeek8a}
\title{Week 8a Answer Function}
\usage{
answerWeek8a(type = "gof")
}
\arguments{
\item{type}{The type of chi-squared test to simulate}
}
\description{
This function computes the answer to the first question of Week 8's practice script
}
