% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerWeek8b.R
\name{answerWeek8b}
\alias{answerWeek8b}
\title{Week 8b Answer Function}
\usage{
answerWeek8b(data, stat)
}
\arguments{
\item{data}{The data from which you will bootstrap}

\item{stat}{The statistic you will bootstrap}
}
\description{
This function computes the answer to the second (really third) question of Week 8's practice script
}
